+++ lib/infospect.dart
@@ -1,6 +1,8 @@
 export 'package:infospect/features/invoker/infospect_desktop_invoker.dart';
 export 'package:infospect/features/invoker/infospect_invoker.dart';
+export 'package:infospect/features/logger/models/infospect_log.dart';
 export 'package:infospect/helpers/infospect_helper.dart';
+export 'package:infospect/infospect.dart';
 export 'package:infospect/styling/themes/infospect_theme.dart';
 // common widgets
 export 'package:infospect/utils/common_widgets/action_widget.dart';