+++ lib/utils/infospect_util.dart
@@ -2,6 +2,10 @@ import 'dart:convert';
 import 'dart:developer' as dev;
 import 'dart:io';
 
+import 'package:infospect/features/logger/models/infospect_log.dart';
+import 'package:infospect/infospect.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+
 class InfospectUtil {
   static bool get isDesktop =>
       (Platform.isLinux || Platform.isWindows || Platform.isMacOS);
@@ -67,5 +71,25 @@ class InfospectUtil {
     } catch (exception) {
       return 'Failed$body';
     }
+  }
+
+  static Future<void> addAppLaunchLog() async {
+    final StringBuffer stringBuffer = StringBuffer();
+    final packageInfo = await PackageInfo.fromPlatform();
+    stringBuffer.write("App name:  ${packageInfo.appName}\n");
+    stringBuffer.write("Package: ${packageInfo.packageName}\n");
+    stringBuffer.write("Version: ${packageInfo.version}\n");
+    stringBuffer.write("Build number: ${packageInfo.buildNumber}\n");
+    stringBuffer.write("Started at: ${DateTime.now().toIso8601String()}\n");
+    stringBuffer.write("\n");
+
+    Infospect.instance.addLog(
+      InfospectLog(
+        message: 'APP STARTED',
+        stackTrace: StackTrace.fromString(
+          stringBuffer.toString(),
+        ),
+      ),
+    );
   }
 }